Fix error in batch mutation updater for AMMR
The wrong indices were returned. This was caught by expanding the tests
of the batch leaf mutation updater method.

+++ src/util_types/mmr/mmr_trait.rs
@@ -17,12 +17,13 @@ where
     /// will end up in a broken state.
     fn mutate_leaf(&mut self, old_membership_proof: &MembershipProof<H>, new_leaf: &H::Digest);
 
-    // Batch mutate an MMR while updating a list of membership proofs
+    // Batch mutate an MMR while updating a list of membership proofs. Returns the indices into
+    // the list of membership proofs that where given as argument.
     fn batch_mutate_leaf_and_update_mps(
         &mut self,
         membership_proofs: &mut Vec<MembershipProof<H>>,
         mutation_data: Vec<(MembershipProof<H>, H::Digest)>,
-    ) -> Vec<u128>;
+    ) -> Vec<usize>;
 
     // Returns true if a list of leaf mutations and a list of appends results in the expected
     // `new_peaks`.
